Modify representation of mapping to None.
I feel that mapping from

```
- from: blast+
  to: blast
```

should be the same as

```
- from:
   name: blast+
   version: Null
  to: blast
```

since it should just be syntatic sugar. And I think that second version reads like it should by default apply to all versions.

I feel therefore it is probably better to explicitly require stating that a mapping only applies to unversioned requirements. This commit does that by introducing the syntax:

```
- from:
    name: blast+
    unversioned: true
  to:
    name: blast
    version: 1.3.4
```

It feels better to have strongly typed entities like this than to introduce magic constants such as:

```
- from:
   name: blast+
   version: __unversioned__
  to:
   name: blast
   version: 1.3.4
```

which is another route this could have gone I guess.