Remove `replace_existing` option for data managers
If a data manager is defined twice for the same data manager id in the
shed_tool_conf.xml the first data manager will be loaded. This is not a
problem when updating a data manager through the toolshed, because the
outdated data manager will be removed from the tool cache and a new
toolbox is being built.