Fix tool submission bursting ?
Problem is that by default we use (as we should) scoped sessions,
so in different threads you're guaranteed to get an independent session.

We precreate the dataset collections before we get to the bursting part,
so when we create the jobs in a thread we have some objects coming form a different
session, and that doesn't work.

I tried using the same session across the burst threads, but that leads
to simultaneous `session.add` and `session.flush` calls. I tried
protecting flushes using a lock, but even with this in place I see
```
InvalidRequestError: This session is in 'prepared' state; no further SQL can be emitted within this transaction.
```
So ultimately I think that's just not a pattern that works well with
sqlalchemy. For faster return times for large mapped-over collections
we should maybe consider passing the whole request blob over to the
job handler and let it create the job ?

To have something to refer to for the client we could return a token
that allows the client to track the job creation process (or pre-create
a Job and return the Job id?).