Use singularity build with biocontainers
This makes it possible to pull a container from
quay.io/biocontainers and to convert it to singularity format
for later execution (via the CachedMulledSingularityContainerResolver).

This is nicely integrated with the DependencyManager view, so
you can (as for conda or docker) do:
```
requests.post(galaxy_url+'/api/tools/deseq2/dependencies', {'key': API_KEY, 'resolver_type': 'mulled_singularity'})
```

This is quite lightweight, the resulting image is significantly smaller
than the conda env (since it is compressed), and the install process
is much faster.