source /usr/local/etc/conda/activate.d/*.sh files
This is needed for containers built from Conda packages that populate
various environment variables in their activate.d scripts.
One example for which this is needed is augustus.

It would probably be better to set and/or record these variables when
building the container, but I think this should cover what is needed
in the IUC (augustus_training only, feelnc, sina).
This should bring biocontainers closer to how one would work with
Conda packages and should benefit other tool maintainers that
haven't started using these workarounds.