Copy working/ and outputs/ before starting job script
This seems to be the most robust way to make sure that resubmission by
the DRM will start with proper configfiles and pre-populated outputs
(OutputParameterJSONTool).
The cost should be minimal as most of the time these are empty
directories.
This would also help with https://github.com/galaxyproject/galaxy/pull/9236#issuecomment-576050103