Avoid flushing in Hierarchical object store
where possible. It seems that whenever we interact with an object
stores' create or update_from_file method we follow up with a flush
anyway. So I think it is actually safe to remove _create_object_in_session
as in 81a3d045dc9ee844bc783fad26a90bf78afd3fdb.