Fix hg_repo path handling in shed_index.py
Entries are absolute paths anyway (at least since 20.05),
and all paths on the main toolshed and test toolshed are absolute.
This only worked for us because os.path.join does this:
```
Signature: os.path.join(a, *p)
Docstring:
Join two or more pathname components, inserting '/' as needed.
If any component is an absolute path, all previous path components
will be discarded.  An empty last part will result in a path that
ends with a separator.
```