Unify getting user by email or username from database
Note that handling the case of more than 1 record returned is different
between the 2 original methods.

-  The regular login retrieves just the
first matching record with `.first()` while `get_api_key` raises an
Exception. I've chosen to not raise an Exception since we don't do this
with regular logins and this targets 20.09. We could experiment with
raising an exception in dev, since that situation should hopefully not
be possible.

- The baseauth method did require exact email capitalization (which was
 a problem for some of Jen's account on usegalaxy.org and for users with
particular LDAP setups, xref: https://github.com/galaxyproject/galaxy/issues/8602)