Replace StreamBall, eliminate tarballs
The way StreamBall writes to the wsgi response just doesn't work for
ASGI frameworks. The zipseeker method seems superior in that it creates
a deflated zip archive, doesn't need a tempfile and makes it easy to set
the content-length header.  To get this to work I had to change the
flatten method to generate the fist item to yield, and then set the
response headers.