Remove questionable expunge_all
This fixes tests in test/integration/test_pulsar_embedded.py, and it's
not quite clear how that would fix "reverting metadata". My guess is
that there was a race condition, but chances are that there is no race
anymore. If it is actually needed this should go upstream of creating
the JobWrapper object.