Implement rough mapping between edam formats and datatypes
This adds the cwl_format parameter to the upload tool and exposes the
cwl_formats field in the details view of the HDA API response, which
contains a list of URIs for the format. This also provides the
edam_format from the datatype annotation as the format field of the cwl
job input description and uses the format field within the cwl job
outputs to write out a gaalxy.json file that contains the cwl-format,
which we then use to pick a datatype (if we find one).

Not sure this is a good idea, in practice this will fall apart for
compressed outputs at least, for which there is no distinction in edam.
If that's the only problem we could probably work around the issue, but
there could be more problems. It is kind of cool and maybe necessary for
interoperability between cwl tool outputs and galaxy tool inputs
(assuming users won't want to change dataypes), but maybe we should just always
stick the edam format into the hda metadata or into a separate hda
column and keep these two concepts separate.

We should maybe extend the edam format use a bit more. This could be
a cool way for tools to use user-defined datatypes that Galaxy doesn't
(yet) know about.