Fix potential thread race in concurrent job limit check
Nate observed a bunch of jobs surpassing limits.
The monitor method is threaded (using the `workers` setting, 4 on
usegalaxy.org). It is possible that where the logic is:
```
if self.count is None:
    self.count = {}
```
Thread 1 might see None, Thread 2 sets the count, Thread 1 sets the count to {}.
It's not striaghtforward to proof, but I think each thread should have
its own cache to work on, hence the contextvars approach.
If it doesn't fix the issue, at least this seems lke the right thing to
do anyway.